modifying unique filter to use doctests
