+++ src/nunavut/lang/c.py
@@ -15,6 +15,7 @@ import typing
 
 import pydsdl
 
+from . import _UniqueNameGenerator
 from .. import SupportsTemplateEnv, templateEnvironmentFilter
 
 # Taken from https://en.cppreference.com/w/c/keyword
@@ -479,27 +480,34 @@ def filter_to_template_unique_name(env: SupportsTemplateEnv, base_token: str) ->
         that the generated name does not conflict with a name generated by
         another means.
 
-    Example::
+    .. invisible-code-block: python
 
-        {{ "foo" | c.to_template_unique_name }}
-        {{ "Foo" | c.to_template_unique_name }}
-        {{ "fOO" | c.to_template_unique_name }}
-        {{ "bar" | c.to_template_unique_name }}
-        {{ "i like coffee" | c.to_template_unique_name }}
+        from nunavut.lang.c import filter_to_template_unique_name
 
-    Results Example::
+    .. code-block:: python
 
-        # These are the likely results but the specific token
-        # generated is not strongly specified.
-        _foo0_
-        _foo1_           # Because '_[A-Z]' is reserved in C this filter will
-                         # lowercase the character after the leading '_'.
-        _fOO0_           # Only the first character lower-casing is enforced.
-                         # The rest of the base token is left un modified.
-        _bar0_
-        _i like coffee0_ # Note that this is not a valid C identifier.
-                         # This filter does _not_ lex the base_token argument
-                         # beyond ensuring the first letter is not [A-Z].
+        # Given
+        template  = '{{ "foo" | to_template_unique_name }},{{ "Foo" | to_template_unique_name }},'
+        template += '{{ "fOO" | to_template_unique_name }}'
+
+        # then
+        rendered = '_foo0_,_foo1_,_fOO0_'
+
+    .. invisible-code-block: python
+
+        jinja_filter_tester(filter_to_template_unique_name, template, rendered)
+
+    .. code-block:: python
+
+        # Given
+        template = '{{ "i like coffee" | to_template_unique_name }}'
+
+        # then
+        rendered = '_i like coffee0_'
+
+    .. invisible-code-block: python
+
+        jinja_filter_tester(filter_to_template_unique_name, template, rendered)
 
 
     :param str base_token: A token to include in the base name.
@@ -511,4 +519,4 @@ def filter_to_template_unique_name(env: SupportsTemplateEnv, base_token: str) ->
     else:
         adj_base_token = base_token
 
-    return env.globals['_unique_name_generator']('c', adj_base_token, '_', '_')  # type: ignore
+    return _UniqueNameGenerator.ensure_generator_in_globals(env.globals)('c', adj_base_token, '_', '_')
