+++ src/nunavut/lang/py.py
@@ -14,6 +14,7 @@ import pydsdl
 
 from .. import SupportsTemplateEnv, templateEnvironmentFilter
 from .c import VariableNameEncoder
+from . import _UniqueNameGenerator
 
 
 @templateEnvironmentFilter
@@ -36,29 +37,40 @@ def filter_to_template_unique_name(env: SupportsTemplateEnv, base_token: str) ->
         that the generated name does not conflict with a name generated by
         another means.
 
-    Example::
+    .. invisible-code-block: python
 
-        {{ "foo" | py.to_template_unique_name }}
-        {{ "foo" | py.to_template_unique_name }}
-        {{ "bar" | py.to_template_unique_name }}
-        {{ "i like coffee" | py.to_template_unique_name }}
+        from nunavut.lang.py import filter_to_template_unique_name
 
-    Results Example::
+    .. code-block:: python
 
-        # These are the likely results but the specific token
-        # generated is not strongly specified.
-        _foo0_
-        _foo1_
-        _bar0_
-        _i like coffee0_ # Note that this is not a valid python identifier.
-                         # This filter does _not_ lex the base_token argument.
+        # Given
+        template  = '{{ "f" | to_template_unique_name }},{{ "f" | to_template_unique_name }},'
+        template += '{{ "f" | to_template_unique_name }},{{ "bar" | to_template_unique_name }}'
 
+        # then
+        rendered = '_f0_,_f1_,_f2_,_bar0_'
+
+    .. invisible-code-block: python
+
+        jinja_filter_tester(filter_to_template_unique_name, template, rendered)
+
+    .. code-block:: python
+
+        # Given
+        template = '{{ "i like coffee" | to_template_unique_name }}'
+
+        # then
+        rendered = '_i like coffee0_'
+
+    .. invisible-code-block: python
+
+        jinja_filter_tester(filter_to_template_unique_name, template, rendered)
 
     :param str base_token: A token to include in the base name.
     :returns: A name that is likely to be valid python identifier and is likely to
         be unique within the file generated by the current template.
     """
-    return env.globals['_unique_name_generator']('py', base_token, '_', '_')  # type: ignore
+    return _UniqueNameGenerator.ensure_generator_in_globals(env.globals)('py', base_token, '_', '_')
 
 
 PYTHON_RESERVED_IDENTIFIERS = frozenset(map(str, list(keyword.kwlist) + dir(builtins)))  # type: typing.FrozenSet[str]
